services/clawtrace-payment/requirements.txt
fastapi==0.118.2
uvicorn[standard]==0.34.0
pydantic==2.12.3
pydantic-settings==2.11.0
asyncpg==0.30.0
python-multipart==0.0.20
stripe>=13.0.0
httpx>=0.27.0
python-json-logger>=3.0.0
azure-communication-email>=1.0.0
azure-identity>=1.25.0
azure-storage-blob>=12.27.0
